Update black configuration
Allow line lengths of 100 for black and remove flake8's config on
length. flake8 is already configured to ignore line length violations.

Reasoning of 100:

A line length of 88 is the default which is 8 more than PEP-8
recommendations. This comes from the default recommendation plus two
indents worth which amount to a class and method. 100 allows for three
more indents. The goal here is not to encourage deeper nesting but to
encourage more verbose naming without trying to trim those few
characteres.

Overall expectation is to keep using a visual guide of 88 characters.