End of year clean up (#177)
* Update pre-commit-config

Version update (autoupdate)

Remove hooks `debug-statements` and `mixed-line-endings` as rarely used.

* Remove development requirements

Both `black` and `flake8` are enforced by pre-commit in the CI and are
not requirements for developing.

`mypy` remains here for support of editor setups that have extensions to
use mypy while working but require the library to be installed.

* Use venv to ensure pip is updated

Use the venv `--update-deps` flag to ensure pip is updated when a new
venv is created.

* Update black configuration

Allow line lengths of 100 for black and remove flake8's config on
length. flake8 is already configured to ignore line length violations.

Reasoning of 100:

A line length of 88 is the default which is 8 more than PEP-8
recommendations. This comes from the default recommendation plus two
indents worth which amount to a class and method. 100 allows for three
more indents. The goal here is not to encourage deeper nesting but to
encourage more verbose naming without trying to trim those few
characteres.

Overall expectation is to keep using a visual guide of 88 characters.

* Update github action versions

Adds the flag `include-hidden-files: true` to the upload artifact action
as behavior was changed in 4.4.1.