Correctly use a constraints file
Turns out I've been using the constraint file incorrectly for years. I
love learning moments like this!

The `constraints.txt` file **is** the lockfile for the project. It
contains all the versions that *should* be used *if* they are needed. Is
is not the source of **what** should be installed, only what version.

So, instead of having N different `requirements*.in` files with N
different `requirements*.txt` generated from `pip-compile`; we have a
single `constraints.txt` that all of our `requirements*.txt` files
reference. Glorious.

This solves for the scaling problem that I ran into when, for example,
`requirements-doc.txt` had a version conflicct with something in
`requirements-test.txt`. Initially, I thought to somehow cross-reference
all the `.txt` files with themselves in some horrible chained constraint
system. That worked about as well as it sounds.

Note, I have to remove the dynamic dependencies here from the
`pyproject.toml`. That's fine. Eventually, I'll be moving to `uv` which
simplifies most of this juggling. For now, however, I want to understand
the underlying machanics before abstracting them away.