Correct nox sessions for CI actions
Okay, this is the ugly part of the path I'm taking. It is now required
to install the correct requirement*.txt files for each session.
Previously it was a simple `.[test]` reference when installing the
package.

This too shall pass.

Already thinking about pivoting into listing all requirements in the
`pyproject.toml`. With dependency groups being supported this should be
amazing. It will hindge on `pip-compile` being able to extract from
dependency groups. That's the next PR!