Refactor running tests with nox
Instead of having a session that handles a large matrix of python
versions, use a session that leverages the default python. This reduces
the overhead of maintaining the `noxfile.py`.

The replacement session has an optional posarg `partial-coverage` which
will produce parallel-mode coverage files. This creates drop-in
replacement behavior for the GitHub action which runs the tests on its
own matrix of python versions and operating systems.

`coverage_combine` moves to a CI only session.