Refactoring noxfile for testing (#189)
* Use nox.options instead of global variable

Move the declaration of the backend to a `nox.options` declaration
instead of passing an extra variable to each `@session`.

* Refactor running tests with nox

Instead of having a session that handles a large matrix of python
versions, use a session that leverages the default python. This reduces
the overhead of maintaining the `noxfile.py`.

The replacement session has an optional posarg `partial-coverage` which
will produce parallel-mode coverage files. This creates drop-in
replacement behavior for the GitHub action which runs the tests on its
own matrix of python versions and operating systems.

`coverage_combine` moves to a CI only session.

* Remove cache: pip option

This is having a negative impact to the CI. Will investigate later