Create lint session for nox (#190)
* Create lint session for nox

Move black, flake8, and isort actions into the `noxfile.py` alongside
the existing mypy session. This will give the dev control over when
these checks are run, allowing pre-commit to be optional in the local
environment.

There is a speed drawback as pre-commit runs on edited files only and
nox will run on all file. The pre-commit config will remain and is
always an option locally and in pre-commit.ci.

* Update CI to use nox lint session

* Remove black config

* Add verbose flag to isort run