Pass python version to uv
This change reverts the use of `python=False` in all nox sessions.
Sessions for tests, linting, and formatting will use nox virtual
environments. In these sessions the location of the virtual environment
will be passed to `uv run` and `uv sync`. This means uv will use the
python version that nox is executed with.

This change is needed because nox correctly masks all environment
variables when starting a `session.run` or `session.run_install`. This
is ideal to maintain repeatable behavior. This is not ideal when
attempting to use `UV_PYTHON` to control which version is being run.

Enforcing that uv use the python version that nox is executed with is
the least surprising behavior. Even if this ignores the
`.python-version` file in the repo.