Adjust noxfile to play nice with uvx and tools
Per the official documentation of uv, `uv tool run` or `uvx` purposely
ignore configuration files that select the python version to use.

This leads to complications when using nox, not wanting nox as a
required dev dependency, and wanting to leverage `uvx nox` or using
`nox` as an installed tool.

Solution: Reapply the `python=False` arguements for all sessions. For
the `test` session which is expected to be run on multiple versions;
check for the existance of `UV_PYTHON` and add that requirement to the
cli flags as needed. For all other sessions, default to the
`.python-version` pinned interpreter.